解决优化算法时间筛选不支持小数的问题

+++ src/SmartContorl/GeneticAlgorithmUI.cpp
@@ -482,9 +482,9 @@ QString GeneticAlgorithmUI::replaceVariate()
 	QString config = "";
 	temp = QString("observeName = \"%1\";\n").arg(this->ui->lineEditName->text());
 	config += temp;
-	temp = QString("maxTime = %1;\n").arg(this->ui->lineEditMaxTime->text().toInt());
+	temp = QString("maxTime = %1;\n").arg(this->ui->lineEditMaxTime->text().toDouble());
 	config += temp;
-	temp = QString("miniTime = %1;\n").arg(this->ui->lineEditMiniTime->text().toInt());
+	temp = QString("miniTime = %1;\n").arg(this->ui->lineEditMiniTime->text().toDouble());
 	config += temp;
 	temp = QString("excpectF = %1;\n").arg(this->ui->lineEditMaxF->text().toLongLong());
 	config += temp;
